requirements.txt
python-telegram-bot==13.15
APScheduler==3.6.3
python-dotenv==1.0.0
pytest==7.4.0
